Set upenvironment for user authentication tests
